requirements.txt
bcrypt==4.1.2
blinker==1.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
pandas==2.2.1
paramiko==3.4.0
pycparser==2.21
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==4.6.2
pymongo-ssh==0.0.15
PyNaCl==1.5.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
sshtunnel==0.4.0
systemd-python==234
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
wadllib==1.3.6
zipp==1.0.0
